requirements.txt
Django==3.0.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-cors-middleware==1.5.0
django-localflavor==2.2
django-money==1.0
Markdown==3.2.1
django-bootstrap4==1.1.1
django-tables2==2.2.1
django-filter==2.2.0
django-markup==1.4.1
docutils==0.16
gunicorn==20.0.4

